Rigging/skinning progress

+++ ShooterGL/ShooterGL/src/Renderables/Mesh.h
@@ -25,7 +25,7 @@ struct Vertex {
 	glm::vec2 TexCoords;
 	glm::vec3 Tangent;
 	glm::vec3 Bitangent;
-	float BoneID[4];
+	int BoneID[4];
 	float WeightValue[4];
 };
